feat: fixed bugs with ai bots, fix: negamaxagent, changed directory for bots
