+++ core/src/main/java/io/deeplay/igorAI/ExpectimaxBot.java
@@ -1,19 +1,20 @@
-package io.deeplay.model.player;
+package io.deeplay.igorAI;
 
+import io.deeplay.igorAI.ai_agent.ExpectimaxAgent;
 import io.deeplay.domain.Color;
-import io.deeplay.ai_agent.MinimaxAgent;
 import io.deeplay.model.Board;
 import io.deeplay.model.move.Move;
+import io.deeplay.model.player.Bot;
 import io.deeplay.service.IUserCommunication;
 
-public class MinimaxBot extends Bot {
+public class ExpectimaxBot extends Bot {
     private static final int DEPTH = 3;
-    public MinimaxBot(Color color, int difficultyLevel, IUserCommunication iUserCommunication) {
+    public ExpectimaxBot(Color color, int difficultyLevel, IUserCommunication iUserCommunication) {
         super(color, difficultyLevel, iUserCommunication);
     }
 
     @Override
     public Move getMove(Board board, Color currentColor) {
-        return new MinimaxAgent().getBestMove(board, DEPTH, Double.MIN_VALUE, Double.MAX_VALUE, getColor());
+        return new ExpectimaxAgent().getBestMove(board, DEPTH, getColor());
     }
 }
