+++ core/src/main/java/io/deeplay/igorAI/ai_agent/MinimaxAgent.java
@@ -1,4 +1,4 @@
-package io.deeplay.ai_agent;
+package io.deeplay.igorAI.ai_agent;
 
 import io.deeplay.domain.Color;
 import io.deeplay.engine.GameState;
@@ -9,34 +9,34 @@ import io.deeplay.model.piece.*;
 import io.deeplay.service.BoardUtil;
 import lombok.Setter;
 
-import java.util.List;
+import java.util.*;
 
-public class ExpectimaxAgent extends AbstractAiAgent {
+public class MinimaxAgent extends AbstractAiAgent {
     /**
      * Setter для тестов функции оценки
      */
     @Setter
     private Color maximizingColor;
     @Setter
-    private Color expectingColor;
+    private Color minimizingColor;
 
-    public Move getBestMove(Board board, int depth, Color currentColor) {
+    public Move getBestMove(Board board, int depth, int alpha, int beta, Color currentColor) {
         maximizingColor = currentColor;
-        expectingColor = currentColor.opposite();
-        return (Move) expectimax(board, depth, currentColor, true)[0];
+        minimizingColor = currentColor.opposite();
+        return (Move) minimax(board, depth, alpha, beta, currentColor, true)[0];
     }
 
-    public Object[] expectimax(Board board, int depth, Color currentColor, boolean maximizingPlayer) {
+    public Object[] minimax(Board board, int depth, int alpha, int beta, Color currentColor, boolean maximizingPlayer) {
         if (depth == 0
-                || GameState.isMate(board, currentColor)
                 || GameState.isStaleMate(board, currentColor)
+                || GameState.isMate(board, currentColor)
                 || GameState.drawWithGameWithoutTakingAndAdvancingPawns(board)) {
             return new Object[]{null, calculatePieces(board, currentColor)};
         }
 
         if (maximizingPlayer) {
             List<Move> allPossibleMoves = getAllPossibleMoves(board, maximizingColor);
-            Move currentBestMove = getRandomMove(allPossibleMoves);
+            Move currentBestMove = null;
             int maxEval = Integer.MIN_VALUE;
 
             for (Move move : allPossibleMoves) {
@@ -44,40 +44,51 @@ public class ExpectimaxAgent extends AbstractAiAgent {
                 BoardUtil.duplicateBoard(board).accept(duplicateBoard);
 
                 duplicateBoard.move(move);
-                int currentEval = (int) expectimax(duplicateBoard, depth - 1, currentColor.opposite(), false)[1];
+                int currentEval = (int) minimax(duplicateBoard, depth - 1, alpha, beta, currentColor.opposite(), false)[1];
 
                 if (currentEval > maxEval) {
                     maxEval = currentEval;
                     currentBestMove = move;
                 }
+
+                alpha = Math.max(alpha, currentEval);
+                if (beta <= alpha) {
+                    break;
+                }
             }
             return new Object[]{currentBestMove, maxEval};
         } else {
-            List<Move> allPossibleMoves = getAllPossibleMoves(board, expectingColor);
-            int expectedMinEval = 0;
+            List<Move> allPossibleMoves = getAllPossibleMoves(board, minimizingColor);
+            Move currentBestMove = getRandomMove(allPossibleMoves);
+            int minEval = Integer.MAX_VALUE;
 
             for (Move move : allPossibleMoves) {
                 Board duplicateBoard = new Board();
                 BoardUtil.duplicateBoard(board).accept(duplicateBoard);
 
                 duplicateBoard.move(move);
-                int currentEval = (int) expectimax(duplicateBoard, depth - 1, currentColor.opposite(), true)[1];
-                expectedMinEval += currentEval;
-            }
+                int currentEval = (int) minimax(duplicateBoard, depth - 1, alpha, beta, currentColor.opposite(), true)[1];
 
-            int avgMinEval = expectedMinEval / allPossibleMoves.size();
-            return new Object[]{null, avgMinEval};
+                if (currentEval < minEval) {
+                    minEval = currentEval;
+                    currentBestMove = move;
+                }
+
+                beta = Math.min(beta, currentEval);
+                if (beta <= alpha) {
+                    break;
+                }
+            }
+            return new Object[]{currentBestMove, minEval};
         }
     }
 
     int calculatePieces(Board board, Color currentColor) {
-        if (GameState.isMate(board, currentColor)) {
-            if (currentColor == maximizingColor) {
-                return -80000;
-            } else if (currentColor == expectingColor) {
-                return 80000;
-            }
-        } else if (GameState.drawWithGameWithoutTakingAndAdvancingPawns(board) || GameState.isStaleMate(board, currentColor)) {
+        if (GameState.isMate(board, currentColor))
+            if (currentColor == maximizingColor) return -8000000;
+            else return 8000000;
+
+        if (GameState.drawWithGameWithoutTakingAndAdvancingPawns(board) || GameState.isStaleMate(board, currentColor)) {
             return 0;
         }
 
@@ -92,7 +103,7 @@ public class ExpectimaxAgent extends AbstractAiAgent {
                         else if (curPiece instanceof Rook) finalScore += 50;
                         else if (curPiece instanceof Queen) finalScore += 90;
                         else if (curPiece instanceof King) finalScore += 900;
-                    } else if (curPiece.getColor() == expectingColor){
+                    } else if (curPiece.getColor() == minimizingColor){
                         if (curPiece instanceof Pawn) finalScore -= 10;
                         else if (curPiece instanceof Knight || curPiece instanceof Bishop) finalScore -= 30;
                         else if (curPiece instanceof Rook) finalScore -= 50;
@@ -102,7 +113,7 @@ public class ExpectimaxAgent extends AbstractAiAgent {
                 }
             }
         }
-
+        // TODO тест на то, что на depth = 2 съедает фигуру
         return finalScore;
     }
 }
